Cleanup kernel extensions system
